spark/requirements.txt
pyspark==3.4.1
pandas
numpy
kafka-python
cassandra-driver
